+++ src/components/moleculars/cards/CardTooltip/index.test.tsx
@@ -1,12 +1,10 @@
 import { renderComponent } from "config/testUtils";
 import { expectTextToBeInTheDocument } from "config/testUtils/expects";
-import CardTextImageTooltip from ".";
+import CardTooltip from ".";
 
 describe("CartTextImageTooltip", () => {
   it("should render without error", () => {
-    renderComponent(
-      <CardTextImageTooltip tooltipText="text" tooltipSymbol="!" />,
-    );
+    renderComponent(<CardTooltip tooltipText="text" tooltipSymbol="!" />);
 
     expectTextToBeInTheDocument("CartTextImageTooltip");
   });
